Create a module for expr classification functions
